+++ packages/@leight/table-client/src/table/index.ts
@@ -1,1 +1,2 @@
+export * from "./SourceTable";
 export * from "./Table";