+++ packages/@leight/utils/src/utils/cleanOf.ts
@@ -7,4 +7,3 @@ export const cleanOf = <T>(obj: T, options?: ICleanerOptions): T => cleaner.clea
     nullCleaner: false,
     ...options,
 });
-1;