+++ packages/@leight/prisma/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leight/prisma
 
+## 0.5.67
+
+### Patch Changes
+
+- Version bump
+
 ## 0.5.66
 
 ### Patch Changes