+++ packages/@leight/table-client/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @leight/table-client
 
+## 0.5.21
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/utils@0.5.55
+    - @leight/i18n@0.5.58
+
 ## 0.5.20
 
 ### Patch Changes