+++ packages/@leight/user-server/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @leight/user-server
 
+## 0.5.63
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/prisma@0.5.67
+    - @leight/utils@0.5.55
+    - @leight/user@0.5.61
+
 ## 0.5.62
 
 ### Patch Changes